Show error banner if copy or move fails (#8322)
