+++ src/Files/Filesystem/FilesystemOperations/Helpers/FilesystemHelpers.cs
@@ -43,8 +43,6 @@ namespace Files.Filesystem
 
         private Task<NamedPipeAsAppServiceConnection> ServiceConnection => AppServiceConnectionHelper.Instance;
 
-        private OngoingTasksViewModel OngoingTasksViewModel => App.OngoingTasksViewModel;
-
         #region Helpers Members
 
         private static readonly char[] RestrictedCharacters = new[] { '\\', '/', ':', '*', '?', '"', '<', '>', '|' };
@@ -111,8 +109,6 @@ namespace Files.Filesystem
         {
             source = await source.ToListAsync();
 
-            var sourceDir = PathNormalization.GetParentDir(source.FirstOrDefault()?.Path);
-            PostedStatusBanner banner = null;
             var returnStatus = ReturnResult.InProgress;
 
             var deleteFromRecycleBin = source.Select(item => item.Path).Any(path => recycleBinHelpers.IsPathUnderRecycleBin(path));
@@ -156,27 +152,7 @@ namespace Files.Filesystem
             }
 
             // post the status banner
-            if (permanently)
-            {
-                // deleting items from <x>
-                banner = OngoingTasksViewModel.PostOperationBanner(string.Empty,
-                    string.Format(source.Count() > 1 ? "StatusDeletingItemsDetails_Plural".GetLocalized() : "StatusDeletingItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir),
-                    0,
-                    ReturnResult.InProgress,
-                    FileOperationType.Delete,
-                    new CancellationTokenSource());
-            }
-            else
-            {
-                // "Moving items from <x> to recycle bin"
-                banner = OngoingTasksViewModel.PostOperationBanner(string.Empty,
-                    string.Format(source.Count() > 1 ? "StatusMovingItemsDetails_Plural".GetLocalized() : "StatusMovingItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir, "TheRecycleBin".GetLocalized()),
-                    0,
-                    ReturnResult.InProgress,
-                    FileOperationType.Recycle,
-                    new CancellationTokenSource());
-            }
-
+            var banner = PostBannerHelpers.PostBanner_Delete(source, returnStatus, permanently, false, 0);
             banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
 
             var token = banner.CancellationToken;
@@ -199,164 +175,14 @@ namespace Files.Filesystem
             banner.Remove();
             sw.Stop();
 
-            if (token.IsCancellationRequested)
-            {
-                if (permanently)
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusDeletionCancelled".GetLocalized(),
-                        string.Format(source.Count() > 1 ?
-                            itemsDeleted > 1 ? "StatusDeleteCanceledDetails_Plural".GetLocalized() : "StatusDeleteCanceledDetails_Plural2".GetLocalized()
-                            : "StatusDeleteCanceledDetails_Singular".GetLocalized(), source.Count(), sourceDir, null, itemsDeleted),
-                        0,
-                        ReturnResult.Cancelled,
-                        FileOperationType.Delete);
-                }
-                else
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusRecycleCancelled".GetLocalized(),
-                        string.Format(source.Count() > 1 ?
-                            itemsDeleted > 1 ? "StatusMoveCanceledDetails_Plural".GetLocalized() : "StatusMoveCanceledDetails_Plural2".GetLocalized()
-                            : "StatusMoveCanceledDetails_Singular".GetLocalized(), source.Count(), sourceDir, "TheRecycleBin".GetLocalized(), itemsDeleted),
-                        0,
-                        ReturnResult.Cancelled,
-                        FileOperationType.Recycle);
-                }
-            }
-            else if (returnStatus == ReturnResult.Success)
-            {
-                if (permanently)
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusDeletionComplete".GetLocalized(),
-                        string.Format(source.Count() > 1 ? "StatusDeletedItemsDetails_Plural".GetLocalized() : "StatusDeletedItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir, itemsDeleted),
-                        0,
-                        ReturnResult.Success,
-                        FileOperationType.Delete);
-                }
-                else
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusRecycleComplete".GetLocalized(),
-                        string.Format(source.Count() > 1 ? "StatusMovedItemsDetails_Plural".GetLocalized() : "StatusMovedItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir, "TheRecycleBin".GetLocalized()),
-                        0,
-                        ReturnResult.Success,
-                        FileOperationType.Recycle);
-                }
-            }
-            else
-            {
-                if (permanently)
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusDeletionFailed".GetLocalized(),
-                        string.Format(source.Count() > 1 ? "StatusDeletionFailedDetails_Plural".GetLocalized() : "StatusDeletionFailedDetails_Singular".GetLocalized(), source.Count(), sourceDir),
-                        0,
-                        ReturnResult.Failed,
-                        FileOperationType.Delete);
-                }
-                else
-                {
-                    OngoingTasksViewModel.PostBanner(
-                        "StatusRecycleFailed".GetLocalized(),
-                        string.Format(source.Count() > 1 ? "StatusMoveFailedDetails_Plural".GetLocalized() : "StatusMoveFailedDetails_Singular".GetLocalized(), source.Count(), sourceDir, "TheRecycleBin".GetLocalized()),
-                        0,
-                        ReturnResult.Failed,
-                        FileOperationType.Recycle);
-                }
-            }
+            PostBannerHelpers.PostBanner_Delete(source, returnStatus, permanently, token.IsCancellationRequested, itemsDeleted);
 
             return returnStatus;
         }
 
         public async Task<ReturnResult> DeleteItemAsync(IStorageItemWithPath source, bool showDialog, bool permanently, bool registerHistory)
         {
-            PostedStatusBanner banner;
-            bool deleteFromRecycleBin = recycleBinHelpers.IsPathUnderRecycleBin(source.Path);
-            var canBeSentToBin = !deleteFromRecycleBin && await recycleBinHelpers.HasRecycleBin(source.Path);
-
-            if (!canBeSentToBin)
-            {
-                permanently = true;
-            }
-
-            if (permanently)
-            {
-                banner = OngoingTasksViewModel.PostBanner(string.Empty,
-                associatedInstance.FilesystemViewModel.WorkingDirectory,
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Delete);
-            }
-            else
-            {
-                banner = OngoingTasksViewModel.PostBanner(string.Empty,
-                associatedInstance.FilesystemViewModel.WorkingDirectory,
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Recycle);
-            }
-
-            var returnStatus = ReturnResult.InProgress;
-
-            banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
-
-            if (((!permanently && !canBeSentToBin) || UserSettingsService.PreferencesSettingsService.ShowConfirmDeleteDialog) && showDialog) // Check if the setting to show a confirmation dialog is on
-            {
-                List<FilesystemItemsOperationItemModel> incomingItems = new List<FilesystemItemsOperationItemModel>();
-
-                var srcPath = source.Path;
-                if (recycleBinHelpers.IsPathUnderRecycleBin(srcPath))
-                {
-                    var binItems = associatedInstance.FilesystemViewModel.FilesAndFolders;
-                    var matchingItem = binItems.FirstOrDefault(x => x.ItemPath == srcPath); // Get original file name
-                    incomingItems.Add(new FilesystemItemsOperationItemModel(FilesystemOperationType.Delete, srcPath, null, matchingItem?.ItemName));
-                }
-                else
-                {
-                    incomingItems.Add(new FilesystemItemsOperationItemModel(FilesystemOperationType.Delete, srcPath, null));
-                }
-
-                FilesystemOperationDialog dialog = FilesystemOperationDialogViewModel.GetDialog(new FilesystemItemsOperationDataModel(
-                    FilesystemOperationType.Delete,
-                    false,
-                    canBeSentToBin ? permanently : true,
-                    canBeSentToBin,
-                    incomingItems,
-                    new List<FilesystemItemsOperationItemModel>()));
-
-                ContentDialogResult result = await dialog.TryShowAsync();
-
-                if (result != ContentDialogResult.Primary)
-                {
-                    banner.Remove();
-                    return ReturnResult.Cancelled; // Return if the result isn't delete
-                }
-
-                // Delete selected item if the result is Yes
-                permanently = dialog.ViewModel.PermanentlyDelete;
-            }
-
-            var sw = new Stopwatch();
-            sw.Start();
-
-            IStorageHistory history = await filesystemOperations.DeleteAsync(source, banner.Progress, banner.ErrorCode, permanently, cancellationToken);
-            ((IProgress<float>)banner.Progress).Report(100.0f);
-            await Task.Yield();
-
-            if (!permanently && registerHistory)
-            {
-                App.HistoryWrapper.AddHistory(history);
-            }
-
-            App.JumpList.RemoveFolder(source.Path); // Remove items from jump list
-
-            banner.Remove();
-            sw.Stop();
-
-            PostBannerHelpers.PostBanner_Delete(returnStatus, permanently ? FileOperationType.Delete : FileOperationType.Recycle, sw, associatedInstance);
-            return returnStatus;
+            return await DeleteItemsAsync(source.CreateEnumerable(), showDialog, permanently, registerHistory);
         }
 
         public async Task<ReturnResult> DeleteItemsAsync(IEnumerable<IStorageItem> source, bool showDialog, bool permanently, bool registerHistory)
@@ -385,19 +211,7 @@ namespace Files.Filesystem
 
         public async Task<ReturnResult> RestoreItemFromTrashAsync(IStorageItemWithPath source, string destination, bool registerHistory)
         {
-            var returnCode = FileSystemStatusCode.InProgress;
-            var errorCode = new Progress<FileSystemStatusCode>();
-            errorCode.ProgressChanged += (s, e) => returnCode = e;
-
-            IStorageHistory history = await filesystemOperations.RestoreFromTrashAsync(source, destination, null, errorCode, cancellationToken);
-
-            if (registerHistory && !string.IsNullOrWhiteSpace(source.Path))
-            {
-                App.HistoryWrapper.AddHistory(history);
-            }
-
-            await Task.Yield();
-            return returnCode.ToStatus();
+            return await RestoreItemsFromTrashAsync(source.CreateEnumerable(), destination.CreateEnumerable(), registerHistory);
         }
 
         public async Task<ReturnResult> RestoreItemsFromTrashAsync(IEnumerable<IStorageItemWithPath> source, IEnumerable<string> destination, bool registerHistory)
@@ -503,21 +317,13 @@ namespace Files.Filesystem
             source = await source.ToListAsync();
             destination = await destination.ToListAsync();
 
-            var sourceDir = PathNormalization.GetParentDir(source.FirstOrDefault()?.Path);
-            var destinationDir = PathNormalization.GetParentDir(destination.FirstOrDefault());
-
-            PostedStatusBanner banner = OngoingTasksViewModel.PostOperationBanner(
-                string.Empty,
-                string.Format(source.Count() > 1 ? "StatusCopyingItemsDetails_Plural".GetLocalized() : "StatusCopyingItemsDetails_Singular".GetLocalized(), source.Count(), destinationDir),
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Copy, new CancellationTokenSource());
-
-            var token = banner.CancellationToken;
-
             var returnStatus = ReturnResult.InProgress;
+
+            var banner = PostBannerHelpers.PostBanner_Copy(source, destination, returnStatus, false, 0);
             banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
 
+            var token = banner.CancellationToken;
+            
             var (collisions, cancelOperation) = await GetCollision(FilesystemOperationType.Copy, source, destination, showDialog);
 
             if (cancelOperation)
@@ -544,87 +350,14 @@ namespace Files.Filesystem
             banner.Remove();
             sw.Stop();
 
-            if (!token.IsCancellationRequested)
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusCopyComplete".GetLocalized(),
-                    string.Format(source.Count() > 1 ? "StatusCopiedItemsDetails_Plural".GetLocalized() : "StatusCopiedItemsDetails_Singular".GetLocalized(), source.Count(), destinationDir, itemsCopied),
-                    0,
-                    ReturnResult.Success,
-                    FileOperationType.Copy);
-            }
-            else
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusCopyCanceled".GetLocalized(),
-                    string.Format(source.Count() > 1 ?
-                        itemsCopied > 1 ? "StatusCopyCanceledDetails_Plural".GetLocalized() : "StatusCopyCanceledDetails_Plural2".GetLocalized() :
-                        "StatusCopyCanceledDetails_Singular".GetLocalized(), source.Count(), destinationDir, itemsCopied),
-                    0,
-                    ReturnResult.Cancelled,
-                    FileOperationType.Copy);
-            }
+            PostBannerHelpers.PostBanner_Copy(source, destination, returnStatus, token.IsCancellationRequested, itemsCopied);
 
             return returnStatus;
         }
 
         public async Task<ReturnResult> CopyItemAsync(IStorageItemWithPath source, string destination, bool showDialog, bool registerHistory)
         {
-            PostedStatusBanner banner = OngoingTasksViewModel.PostBanner(
-                string.Empty,
-                associatedInstance.FilesystemViewModel.WorkingDirectory,
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Copy);
-
-            var returnStatus = ReturnResult.InProgress;
-            banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
-
-            var (collisions, cancelOperation) = await GetCollision(FilesystemOperationType.Copy, source.CreateList(), destination.CreateList(), showDialog);
-
-            if (cancelOperation)
-            {
-                banner.Remove();
-                return ReturnResult.Cancelled;
-            }
-
-            var sw = new Stopwatch();
-            sw.Start();
-
-            itemManipulationModel?.ClearSelection();
-
-            IStorageHistory history = null;
-            if (collisions.First() != FileNameConflictResolveOptionType.Skip)
-            {
-                history = await filesystemOperations.CopyAsync(source, destination, collisions.First().Convert(), banner.Progress, banner.ErrorCode, cancellationToken);
-                ((IProgress<float>)banner.Progress).Report(100.0f);
-                await Task.Yield();
-            }
-            else
-            {
-                ((IProgress<float>)banner.Progress).Report(100.0f);
-                return ReturnResult.Cancelled;
-            }
-
-            if (registerHistory && !string.IsNullOrWhiteSpace(source.Path))
-            {
-                App.HistoryWrapper.AddHistory(history);
-            }
-
-            banner.Remove();
-            sw.Stop();
-
-            if (sw.Elapsed.TotalSeconds >= 10)
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusCopyComplete".GetLocalized(),
-                    "StatusOperationCompleted".GetLocalized(),
-                    0,
-                    ReturnResult.Success,
-                    FileOperationType.Copy);
-            }
-
-            return returnStatus;
+            return await CopyItemsAsync(source.CreateEnumerable(), destination.CreateEnumerable(), showDialog, registerHistory);
         }
 
         public async Task<ReturnResult> CopyItemsFromClipboard(DataPackageView packageView, string destination, bool showDialog, bool registerHistory)
@@ -724,20 +457,15 @@ namespace Files.Filesystem
             source = await source.ToListAsync();
             destination = await destination.ToListAsync();
 
+            var returnStatus = ReturnResult.InProgress;
+
             var sourceDir = PathNormalization.GetParentDir(source.FirstOrDefault()?.Path);
             var destinationDir = PathNormalization.GetParentDir(destination.FirstOrDefault());
 
-            PostedStatusBanner banner = OngoingTasksViewModel.PostOperationBanner(
-                string.Empty,
-                string.Format(source.Count() > 1 ? "StatusMovingItemsDetails_Plural".GetLocalized() : "StatusMovingItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir, destinationDir),
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Move, new CancellationTokenSource());
+            var banner = PostBannerHelpers.PostBanner_Move(source, destination, returnStatus, false, 0);
+            banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
 
             var token = banner.CancellationToken;
-
-            var returnStatus = ReturnResult.InProgress;
-            banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
 
             var (collisions, cancelOperation) = await GetCollision(FilesystemOperationType.Move, source, destination, showDialog);
 
@@ -767,96 +495,14 @@ namespace Files.Filesystem
             banner.Remove();
             sw.Stop();
 
-            if (!token.IsCancellationRequested)
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusMoveComplete".GetLocalized(),
-                    string.Format(source.Count() > 1 ? "StatusMovedItemsDetails_Plural".GetLocalized() : "StatusMovedItemsDetails_Singular".GetLocalized(), source.Count(), sourceDir, destinationDir, itemsMoved),
-                    0,
-                    ReturnResult.Success,
-                    FileOperationType.Move);
-            }
-            else
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusMoveCanceled".GetLocalized(),
-                    string.Format(source.Count() > 1 ?
-                        itemsMoved > 1 ? "StatusMoveCanceledDetails_Plural".GetLocalized() : "StatusMoveCanceledDetails_Plural2".GetLocalized()
-                        : "StatusMoveCanceledDetails_Singular".GetLocalized(), source.Count(), sourceDir, destinationDir, itemsMoved),
-                    0,
-                    ReturnResult.Cancelled,
-                    FileOperationType.Move);
-            }
+            PostBannerHelpers.PostBanner_Move(source, destination, returnStatus, token.IsCancellationRequested, itemsMoved);
 
             return returnStatus;
         }
 
         public async Task<ReturnResult> MoveItemAsync(IStorageItemWithPath source, string destination, bool showDialog, bool registerHistory)
         {
-            PostedStatusBanner banner = OngoingTasksViewModel.PostBanner(
-                string.Empty,
-                associatedInstance.FilesystemViewModel.WorkingDirectory,
-                0,
-                ReturnResult.InProgress,
-                FileOperationType.Move);
-
-            var returnStatus = ReturnResult.InProgress;
-            banner.ErrorCode.ProgressChanged += (s, e) => returnStatus = e.ToStatus();
-
-            var (collisions, cancelOperation) = await GetCollision(FilesystemOperationType.Move, source.CreateList(), destination.CreateList(), showDialog);
-
-            if (cancelOperation)
-            {
-                banner.Remove();
-                return ReturnResult.Cancelled;
-            }
-
-            if (cancelOperation)
-            {
-                banner.Remove();
-                return ReturnResult.Cancelled;
-            }
-
-            var sw = new Stopwatch();
-            sw.Start();
-
-            itemManipulationModel?.ClearSelection();
-
-            IStorageHistory history = null;
-
-            if (collisions.First() != FileNameConflictResolveOptionType.Skip)
-            {
-                history = await filesystemOperations.MoveAsync(source, destination, collisions.First().Convert(), banner.Progress, banner.ErrorCode, cancellationToken);
-                ((IProgress<float>)banner.Progress).Report(100.0f);
-                await Task.Yield();
-            }
-            else
-            {
-                ((IProgress<float>)banner.Progress).Report(100.0f);
-                return ReturnResult.Cancelled;
-            }
-
-            if (registerHistory && !string.IsNullOrWhiteSpace(source.Path))
-            {
-                App.HistoryWrapper.AddHistory(history);
-            }
-
-            App.JumpList.RemoveFolder(source.Path); // Remove items from jump list
-
-            banner.Remove();
-            sw.Stop();
-
-            if (sw.Elapsed.TotalSeconds >= 10)
-            {
-                OngoingTasksViewModel.PostBanner(
-                    "StatusMoveComplete".GetLocalized(),
-                    "StatusOperationCompleted".GetLocalized(),
-                    0,
-                    ReturnResult.Success,
-                    FileOperationType.Move);
-            }
-
-            return returnStatus;
+            return await MoveItemsAsync(source.CreateEnumerable(), destination.CreateEnumerable(), showDialog, registerHistory);
         }
 
         public async Task<ReturnResult> MoveItemsFromClipboard(DataPackageView packageView, string destination, bool showDialog, bool registerHistory)
